exception dump

+++ types.h
@@ -1,4 +1,4 @@
 #pragma once
 
-typedef unsigned long long uint64_t;
+typedef unsigned long uint64_t;
 typedef uint64_t word_t;
